Update ENS manager
