--- src/modules/wallets/components/register-domain-overlay/components/CompleteRegistration/index.js
@@ -1,1 +1,0 @@
-export { default } from './CompleteRegistration';
