--- src/modules/wallets/components/register-domain-overlay/components/DomainTaken/index.js
@@ -1,1 +1,0 @@
-export { default } from './DomainTaken';
