+++ src/modules/wallets/components/register-domain-overlay/components/Step3/Step3.vue
@@ -1,12 +1,14 @@
 <template>
   <div>
-    aaa
+    <complete-registration />
   </div>
 </template>
 
 <script>
+import CompleteRegistration from '../complete-registration/CompleteRegistration';
+
 export default {
-  components: {},
+  components: { CompleteRegistration },
   props: {},
   data() {
     return {};
